# perf(trace): optimize decodeHex function and add benchmarks